users_db.go にデータベースを扱う関数をまとめた
